Added confirm dialog ui
Added function to open confirmation dialog and connected AnimationDetails with the dialog through UIManager. Also renamed GUI to UI or "MenuUI", because there could be some stuff that aren't 100% related to graphics and I can't be bothered to be precise with the definitions so I just named it with a "wider brush" so to speak.