Fixed precision mode
Precision mode seems to work properly now. Also fixed an issue with timestamp for frames not starting from 0 after trying to re-record after a failed attempt with precision on. Fix was adding previous timestamp time with frame's timestamp. Honestly not sure 100% why it works but I speculate that threads and the frame's timestamp being updated inside the frameshandler class might have something to do with it.