Somewhat fixed issues with poses
I got the poses to rendering by remembering to set a parent for them (they weren't in the UI layer). Also got rid of most error messages. Code structure feels bad and the poses move too fast and not timed properly with playback, also first pose is not even a pose.