Fixed(?) calibration issues
Fixed calibration issues by applying scaled coordinates in relative to pelvis. Before the scaled body seemed to move forward and pelvis stayed behind.
Also fixed issues with fingers being longer than usual. I used the scaling factor of the hand instead of the fingers. I noticed that scaling factors to all joints seem to be pretty close to each other except for fingers. Probably not the ideal solution for a unique body type, but unfortunately I don't have time to come up with a solution for that.