Implemented UI
Implemented several UI components to the game scene. It now has a window about calibration, finish screen that tells the overall accuracy, score text on each pose, and updated live body tracking display.

Also implemented simple difficulty select on main menu.

Because of time constraints lots of stuff related to game scene is pretty rough and hard to read, but at least it works for now. This might be the last time I'll be able to update this code in a while.